+++ nekoyume/Assets/_Scripts/GameConfig.cs
@@ -13,11 +13,11 @@ namespace Nekoyume
         public const int DefaultAvatarWeaponId = 10100000;
         public const int DefaultAvatarArmorId = 10200000;
         
-        public const float CombinationValueP1 = 3f; // 30f;
-        public const float CombinationValueP2 = 1f; // 1.2f;
-        public const float CombinationValueL1 = 10f;
-        public const float CombinationValueL2 = 1f;
-        public const float CombinationValueR1 = 1.5f; // 2f;
+        public const decimal CombinationValueP1 = 3m; // 30f;
+        public const decimal CombinationValueP2 = 1m; // 1.2f;
+        public const decimal CombinationValueL1 = 10m;
+        public const decimal CombinationValueL2 = 1m;
+        public const decimal CombinationValueR1 = 1.5m; // 2f;
         public const int CombinationDefaultFoodId = 200000;
         
         #endregion
